+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * `@mixin di()` en `@mixin di-variant()` uit @dso-toolkit/core verwijderen ([#1752](https://github.com/dso-toolkit/dso-toolkit/issues/1752))
 * `@mixin di()` verhuizen van `packages/sources` naar `packages/css` ([#1804](https://github.com/dso-toolkit/dso-toolkit/issues/1804))
 * **core + sources:** Pagination: Werkvorm afstemmen met beschikbare ruimte ([#1772](https://github.com/dso-toolkit/dso-toolkit/issues/1772))
+* **core + sources** Ozon content <Figuur> Titel, bijschrift, en bron mapping ([#1730](https://github.com/dso-toolkit/dso-toolkit/issues/1730))
 
 ## 45.0.1